Refactor all of the classes into subpackages
No functional changes but refactor all the methods into submodules
which makes it easier for me to image refactoring the code further
