+++ datargs/dispatch/__init__.py
@@ -1,0 +1,2 @@
+import datargs.dispatch.concrete
+from .registry import TypeRegistry
